# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swimlane Studio API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Workspaces",
        "operationId": "workspaces_api_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Workspace"
                  },
                  "type": "array",
                  "title": "Response Workspaces Api Workspaces Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Register",
        "operationId": "register_api_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "History",
        "operationId": "history_api_tickets__ticket_id__runs_get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  },
                  "title": "Response History Api Tickets  Ticket Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Activity",
        "operationId": "activity_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunEvent"
                  },
                  "title": "Response Activity Api Runs  Run Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/cancel": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_tickets__ticket_id__cancel_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/retry": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Retry",
        "operationId": "retry_api_tickets__ticket_id__retry_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Events",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Groups",
        "operationId": "groups_api_groups_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  },
                  "title": "Response Groups Api Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Create",
        "operationId": "create_api_groups_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{id}": {
      "patch": {
        "tags": [
          "Groups"
        ],
        "summary": "Update",
        "operationId": "update_api_groups__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Delete",
        "operationId": "delete_api_groups__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{id}/group": {
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Assign",
        "operationId": "assign_api_tickets__id__group_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupAssignment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{id}/run": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Run Group",
        "operationId": "run_group_api_groups__id__run_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  },
                  "title": "Response Run Group Api Groups  Id  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChangedFile": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "additions": {
            "type": "integer",
            "title": "Additions"
          },
          "deletions": {
            "type": "integer",
            "title": "Deletions"
          },
          "diff": {
            "type": "string",
            "title": "Diff"
          }
        },
        "type": "object",
        "required": [
          "path",
          "additions",
          "deletions",
          "diff"
        ],
        "title": "ChangedFile"
      },
...
      "Draft": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "prompt": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt"
          },
          "provider": {
            "type": "string",
            "enum": [
              "demo",
              "codex",
              "claude"
            ],
            "title": "Provider",
            "default": "demo"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Images"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": ""
          },
          "permission": {
            "type": "string",
            "enum": [
              "read-only",
              "workspace-write"
            ],
            "title": "Permission",
            "default": "read-only"
          },
          "allow_tests": {
            "type": "boolean",
            "title": "Allow Tests",
            "default": false
          },
          "group_id": {
            "type": "string",
            "maxLength": 80,
            "title": "Group Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "prompt"
        ],
        "title": "Draft"
      },
      "Group": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "color": {
            "type": "string",
            "pattern": "^(orange|blue|green|amber|plum|#[0-9a-fA-F]{6})$",
            "title": "Color",
            "default": "orange"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": "default"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "Group"
      },
      "GroupAssignment": {
        "properties": {
          "group_id": {
            "type": "string",
            "maxLength": 80,
            "title": "Group Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "GroupAssignment"
      },
      "GroupInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "color": {
            "type": "string",
            "pattern": "^(orange|blue|green|amber|plum|#[0-9a-fA-F]{6})$",
            "title": "Color",
            "default": "orange"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Run": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ticket_id": {
            "type": "string",
            "title": "Ticket Id"
          },
          "snapshot": {
            "$ref": "#/components/schemas/Draft"
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "queued",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "default": "queued"
          },
          "group_run_id": {
            "type": "string",
            "title": "Group Run Id",
            "default": ""
          },
          "previous_run_id": {
            "type": "string",
            "title": "Previous Run Id",
            "default": ""
          },
          "base_commit": {
            "type": "string",
            "title": "Base Commit",
            "default": ""
          },
          "workspace_path": {
            "type": "string",
            "title": "Workspace Path",
            "default": ""
          },
          "worktree": {
            "type": "string",
            "title": "Worktree",
            "default": ""
          },
          "test_preset": {
            "type": "string",
            "enum": [
              "none",
              "vitest",
              "pytest"
            ],
            "title": "Test Preset",
            "default": "none"
          },
          "test_directory": {
            "type": "string",
            "title": "Test Directory",
            "default": "."
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "type": "string",
            "title": "Started At",
            "default": ""
          },
          "finished_at": {
            "type": "string",
            "title": "Finished At",
            "default": ""
          },
          "response": {
            "type": "string",
            "title": "Response",
            "default": ""
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ChangedFile"
            },
            "type": "array",
            "title": "Files"
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          },
          "tests": {
            "type": "string",
            "title": "Tests",
            "default": "not-requested"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticket_id",
          "snapshot",
          "created_at"
        ],
        "title": "Run"
      },
      "RunEvent": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "ticket_id": {
            "type": "string",
            "title": "Ticket Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "ticket_id",
          "type",
          "message",
          "created_at"
        ],
        "title": "RunEvent"
      },
      "Ticket": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "prompt": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt"
          },
          "provider": {
            "type": "string",
            "enum": [
              "demo",
              "codex",
              "claude"
            ],
            "title": "Provider",
            "default": "demo"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Images"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": ""
          },
          "permission": {
            "type": "string",
            "enum": [
              "read-only",
              "workspace-write"
            ],
            "title": "Permission",
            "default": "read-only"
          },
          "allow_tests": {
            "type": "boolean",
            "title": "Allow Tests",
            "default": false
          },
          "group_id": {
            "type": "string",
            "maxLength": 80,
            "title": "Group Id",
            "default": ""
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": "default"
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "queued",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "default": "todo"
          },
          "response": {
            "type": "string",
            "title": "Response",
            "default": ""
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ChangedFile"
            },
            "type": "array",
            "title": "Files"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "queued_at": {
            "type": "string",
            "title": "Queued At",
            "default": ""
          },
          "latest_run_id": {
            "type": "string",
            "title": "Latest Run Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "prompt",
          "id",
          "created_at"
        ],
        "title": "Ticket"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Workspace": {
        "properties": {
          "path": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "test_preset": {
            "type": "string",
            "enum": [
              "none",
              "vitest",
              "pytest"
            ],
            "title": "Test Preset",
            "default": "none"
          },
          "test_directory": {
            "type": "string",
            "maxLength": 500,
            "title": "Test Directory",
            "default": "."
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "path",
          "name",
          "id"
        ],
        "title": "Workspace"
      },
      "WorkspaceInput": {
        "properties": {
          "path": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "test_preset": {
            "type": "string",
            "enum": [
              "none",
              "vitest",
              "pytest"
            ],
            "title": "Test Preset",
            "default": "none"
          },
          "test_directory": {
            "type": "string",
            "maxLength": 500,
            "title": "Test Directory",
            "default": "."
          }
        },
        "type": "object",
        "required": [
          "path",
          "name"
        ],
        "title": "WorkspaceInput"
      }
    }
  },
...
}